+++ src/main/java/prototype/hifi/dnick/service/TestResultService.java
@@ -6,7 +6,7 @@ import prototype.hifi.dnick.model.User;
 import java.util.List;
 
 public interface TestResultService {
-    TestResult saveNewResult(User user, Double points);
+    TestResult saveNewResult(User user, int points);
     void initResultsForUser(User user);
     List<TestResult> getTopThreeResultsForUser(User user);
     int getNumberOfTestsForUser(User user);
